# Speed up the linters