--- a/lints/lint_distribution_point_incomplete.go
+++ b/lints/lint_distribution_point_incomplete.go
@@ -16,6 +16,7 @@ package lints
 
 import (
 	"encoding/asn1"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zcrypto/x509/pkix"
 	"github.com/zmap/zlint/util"