--- a/lints/lint_distribution_point_missing_ldap_or_uri.go
+++ b/lints/lint_distribution_point_missing_ldap_or_uri.go
@@ -7,9 +7,10 @@ When present, DistributionPointName SHOULD include at least one LDAP or HTTP URI
 package lints
 
 import (
+	"strings"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
-	"strings"
 )
 
 type distribNoLDAPorURI struct{}