--- a/lints/lint_dnsname_check_left_label_wildcard.go
+++ b/lints/lint_dnsname_check_left_label_wildcard.go
@@ -1,9 +1,10 @@
 package lints
 
 import (
+	"strings"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
-	"strings"
 )
 
 type DNSNameLeftLabelWildcardCheck struct{}