--- a/lints/lint_dnsname_label_too_long.go
+++ b/lints/lint_dnsname_label_too_long.go
@@ -1,9 +1,10 @@
 package lints
 
 import (
+	"strings"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
-	"strings"
 )
 
 type DNSNameLabelLengthTooLong struct{}