--- a/lints/lint_dnsname_underscore_in_sld.go
+++ b/lints/lint_dnsname_underscore_in_sld.go
@@ -19,7 +19,7 @@ func (l *DNSNameUnderscoreInSLD) CheckApplies(c *x509.Certificate) bool {
 
 func (l *DNSNameUnderscoreInSLD) Execute(c *x509.Certificate) *LintResult {
 	if c.Subject.CommonName != "" {
-		var domainInfo = c.GetParsedSubjectCommonName(false)
+		domainInfo := c.GetParsedSubjectCommonName(false)
 		if domainInfo.ParseError != nil {
 			return &LintResult{Status: NA}
 		}
@@ -28,7 +28,7 @@ func (l *DNSNameUnderscoreInSLD) Execute(c *x509.Certificate) *LintResult {
 		}
 	}
 
-	var parsedSANDNSNames = c.GetParsedDNSNames(false)
+	parsedSANDNSNames := c.GetParsedDNSNames(false)
 	for i := range c.GetParsedDNSNames(false) {
 		if parsedSANDNSNames[i].ParseError != nil {
 			return &LintResult{Status: NA}