--- a/lints/lint_dnsname_underscore_in_trd.go
+++ b/lints/lint_dnsname_underscore_in_trd.go
@@ -17,37 +17,27 @@ func (l *DNSNameUnderscoreInTRD) CheckApplies(c *x509.Certificate) bool {
 	return util.IsSubscriberCert(c) && util.DNSNamesExist(c)
 }
 
-func underscoreInTRD(domain string) (bool, error) {
-	domainName, err := util.ICANNPublicSuffixParse(domain)
-	if err != nil {
-		return true, err
-	}
-	if strings.Contains(domainName.TRD, "_") {
-		return true, nil
-	} else {
-		return false, nil
-	}
-}
-
 func (l *DNSNameUnderscoreInTRD) Execute(c *x509.Certificate) *LintResult {
 	if c.Subject.CommonName != "" {
-		underscoreFound, err := underscoreInTRD(c.Subject.CommonName)
-		if err != nil {
+		var domainInfo = c.GetParsedSubjectCommonName(false)
+		if domainInfo.ParseError != nil {
 			return &LintResult{Status: NA}
 		}
-		if underscoreFound {
+		if strings.Contains(domainInfo.ParsedDomain.TRD, "_") {
 			return &LintResult{Status: Warn}
 		}
 	}
-	for _, dns := range c.DNSNames {
-		underscoreFound, err := underscoreInTRD(dns)
-		if err != nil {
+
+	var parsedSANDNSNames = c.GetParsedDNSNames(false)
+	for i := range c.GetParsedDNSNames(false) {
+		if parsedSANDNSNames[i].ParseError != nil {
 			return &LintResult{Status: NA}
 		}
-		if underscoreFound {
+		if strings.Contains(parsedSANDNSNames[i].ParsedDomain.TRD, "_") {
 			return &LintResult{Status: Warn}
 		}
 	}
+
 	return &LintResult{Status: Pass}
 }