--- a/lints/lint_dnsname_wildcard_only_in_left_label.go
+++ b/lints/lint_dnsname_wildcard_only_in_left_label.go
@@ -1,9 +1,10 @@
 package lints
 
 import (
+	"strings"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
-	"strings"
 )
 
 type DNSNameWildcardOnlyInLeftlabel struct{}