--- a/lints/lint_ext_aia_access_location_missing.go
+++ b/lints/lint_ext_aia_access_location_missing.go
@@ -13,9 +13,10 @@ An authorityInfoAccess extension may include multiple instances of
 package lints
 
 import (
+	"strings"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
-	"strings"
 )
 
 type aiaNoHTTPorLDAP struct{}