--- a/lints/lint_ext_policy_constraints_empty.go
+++ b/lints/lint_ext_policy_constraints_empty.go
@@ -12,6 +12,7 @@ package lints
 
 import (
 	"encoding/asn1"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
 )