--- a/lints/lint_ext_san_uri_not_ia5.go
+++ b/lints/lint_ext_san_uri_not_ia5.go
@@ -7,9 +7,10 @@ stored in the uniformResourceIdentifier (an IA5String).
 package lints
 
 import (
+	"unicode"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
-	"unicode"
 )
 
 type extSANURINotIA5 struct{}