--- a/lints/lint_generalized_time_includes_fraction_seconds.go
+++ b/lints/lint_generalized_time_includes_fraction_seconds.go
@@ -16,6 +16,7 @@ package lints
 
 import (
 	"encoding/asn1"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
 )