--- a/lints/lint_ian_bare_wildcard.go
+++ b/lints/lint_ian_bare_wildcard.go
@@ -3,9 +3,10 @@
 package lints
 
 import (
+	"strings"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
-	"strings"
 )
 
 type brIANBareWildcard struct{}