--- a/lints/lint_ian_dns_name_starts_with_period.go
+++ b/lints/lint_ian_dns_name_starts_with_period.go
@@ -3,9 +3,10 @@
 package lints
 
 import (
+	"strings"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
-	"strings"
 )
 
 type IANDNSPeriod struct{}