--- a/lints/lint_ian_iana_pub_suffix_empty.go
+++ b/lints/lint_ian_iana_pub_suffix_empty.go
@@ -3,9 +3,10 @@
 package lints
 
 import (
+	"strings"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
-	"strings"
 )
 
 type IANPubSuffix struct{}