--- a/lints/lint_idn_dnsname_malformed_unicode.go
+++ b/lints/lint_idn_dnsname_malformed_unicode.go
@@ -1,10 +1,11 @@
 package lints
 
 import (
+	"strings"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
 	"golang.org/x/net/idna"
-	"strings"
 )
 
 type IDNMalformedUnicode struct{}