--- a/lints/lint_idn_dnsname_must_be_nfkc.go
+++ b/lints/lint_idn_dnsname_must_be_nfkc.go
@@ -1,11 +1,12 @@
 package lints
 
 import (
+	"strings"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
 	"golang.org/x/net/idna"
 	"golang.org/x/text/unicode/norm"
-	"strings"
 )
 
 type IDNNotNFKC struct{}