--- a/lints/lint_is_redacted_cert.go
+++ b/lints/lint_is_redacted_cert.go
@@ -1,9 +1,10 @@
 package lints
 
 import (
+	"strings"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
-	"strings"
 )
 
 type DNSNameRedacted struct{}