--- a/lints/lint_rsa_exp_negative.go
+++ b/lints/lint_rsa_exp_negative.go
@@ -4,6 +4,7 @@ package lints
 
 import (
 	"crypto/rsa"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
 )