--- a/lints/lint_rsa_mod_factors_smaller_than_752_bits.go
+++ b/lints/lint_rsa_mod_factors_smaller_than_752_bits.go
@@ -8,6 +8,7 @@ package lints
 
 import (
 	"crypto/rsa"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
 )
@@ -29,6 +30,7 @@ func (l *rsaModSmallFactor) Execute(c *x509.Certificate) *LintResult {
 		return &LintResult{Status: Pass}
 	}
 	return &LintResult{Status: Warn}
+
 }
 
 func init() {