--- a/lints/lint_rsa_no_public_key.go
+++ b/lints/lint_rsa_no_public_key.go
@@ -3,6 +3,7 @@ package lints
 
 import (
 	"crypto/rsa"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
 )