--- a/lints/lint_rsa_public_exponent_not_odd.go
+++ b/lints/lint_rsa_public_exponent_not_odd.go
@@ -8,6 +8,7 @@ package lints
 
 import (
 	"crypto/rsa"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
 )