--- a/lints/lint_sub_cert_is_ca.go
+++ b/lints/lint_sub_cert_is_ca.go
@@ -2,6 +2,7 @@ package lints
 
 import (
 	"encoding/asn1"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
 )