--- a/lints/lint_subject_contains_reserved_ip.go
+++ b/lints/lint_subject_contains_reserved_ip.go
@@ -11,9 +11,10 @@ Address or Internal Name.
 package lints
 
 import (
+	"net"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
-	"net"
 )
 
 type subjectReservedIP struct{}