--- a/lints/lint_subject_not_dn.go
+++ b/lints/lint_subject_not_dn.go
@@ -11,10 +11,11 @@
 package lints
 
 import (
+	"reflect"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zcrypto/x509/pkix"
 	"github.com/zmap/zlint/util"
-	"reflect"
 )
 
 type subjectDN struct{}