--- a/lints/lint_subject_organizational_unit_name_max_length.go
+++ b/lints/lint_subject_organizational_unit_name_max_length.go
@@ -9,9 +9,10 @@ RFC 5280: A.1
 package lints
 
 import (
+	"unicode/utf8"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
-	"unicode/utf8"
 )
 
 type subjectOrganizationalUnitNameMaxLength struct{}