--- a/lints/lint_utc_time_not_in_zulu.go
+++ b/lints/lint_utc_time_not_in_zulu.go
@@ -20,9 +20,10 @@
 package lints
 
 import (
+	"time"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
-	"time"
 )
 
 type utcTimeGMT struct {