--- a/lints/lint_wrong_time_format_pre2050.go
+++ b/lints/lint_wrong_time_format_pre2050.go
@@ -11,9 +11,10 @@ package lints
 
 import (
 	"encoding/asn1"
+	"time"
+
 	"github.com/zmap/zcrypto/x509"
 	"github.com/zmap/zlint/util"
-	"time"
 )
 
 type generalizedPre2050 struct{}