This pull request will speed up the slowest linters. 

* w_dnsname_underscore_in_trd
* e_dnsname_underscore_in_sld
* e_dnsname_hyphen_in_sld
* w_dnsname_wildcard_left_of_public_suffix
* w_san_iana_pub_suffix_empty
* w_rsa_mod_factors_smaller_than_752
* e_dnsname_bad_character_in_label

Plus some linting done by my IDE. 

Benchmark results
```
benchcmp old.benchmarks.txt new.benchmarks.txt
benchmark                      old ns/op     new ns/op     delta
BenchmarkZlint/All_lints-8     33228743      407902        -98.77%

benchmark                      old allocs     new allocs     delta
BenchmarkZlint/All_lints-8     30006          2115           -92.95%

benchmark                      old bytes     new bytes     delta
BenchmarkZlint/All_lints-8     18835646      90210         -99.52%
```

Overall, the benchmarks are ~80x faster. 
If you guys are happy with the changes, I might have a look at some other slow benchmarks. 

* w_subject_dn_leading_whitespace
* w_subject_dn_trailing_whitespace
* w_multiple_subject_rdn
* e_ext_san_dns_not_ia5_string
* e_ext_san_empty_name
* e_dnsname_not_valid_tld
* e_dnsname_contains_bare_iana_suffix
* e_dnsname_wildcard_only_in_left_label
* e_international_dns_name_not_nfkc

Again, speeding them up might require some changes to the zcrypto/x509 to cache some values (tentatively)

I would appreciate any remarks.