@szank 2017-11-20
Hi, 
I just wanted to ping you guys. Did you have a chance to have a look at the updated PR ?